chore(fv-std): Move `fv_std` logic in separate file
